fix: separate xml parser util
